requirements.txt
python-dotenv
requests
svgwrite
